==> 02 Scripts/01 Orders - Data Wrangling & Data Consistency Checks.ipynb <==
# Orders - Data Wrangling and Data Consistency Checks:

1. Importing libraries and dataset
2. Checking for columns, datatype, shape using .info()
3. Removing unneccessary columns using .drop()
4. Renaming columns using .rename()
5. Addressing missing values
6. Addressing duplicates
7. Checking for mixed datatype
8. Changing datatypes to reduce memory usage
9. Performing Descriptive Analysis
10. Exporting wrangled, consistency checked dataframe

## 1. Importing libraries and dataset

# Importing libraries

import pandas as pd
import os

# Accessing EnvFile for path

%run EnvFile.ipynb

# Importing orders.csv to dataframe

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'))

# Checking the head

df_ords.head()

## 2. Checking for columns, datatype, shape using .info()

# Checking the info for columns, datatypes, shape of dataframe

df_ords.info()

#### The shape of df_ords before consistency checks is (3421083, 7) with memory usage of 182.7+ MB.

## 3. Removing unneccessary columns using .drop()

# Removing unnecessary column eval_set

df_ords.drop(columns = ['eval_set'], inplace = True)

## 4. Renaming columns using .rename()

# Renaming columns with appropriate names

df_ords.rename(columns = {'order_dow':'order_day_of_week'}, inplace = True)

## 5. Addressing missing values

# Finding missing values

df_ords.isnull().sum()

# Creating a subset of missing values in df_ords

df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_ords_nan

#### Findings of missing values

The 206209 missing values of days_since_prior_order in df_ords_nan imply that all those are orders from order_number 1 for all the users from 1 to 206209. 
It makes sense because there wouldn't be any orders prior to order_number 1. Hence there is no days since prior order.

#### Addressing missing values

I've chosen not to impute or remove null value since statistics can still be calculated with it and those rows are needed to get input on first orders from all customers. 
The order number 1 act as a flag already resulting in no necessity for separate flag column to be created.

## 6. Addressing duplicates

# Check for duplicates

df_ords[df_ords.duplicated()]

There are no duplicates. Hence no changes to be made.

## 7. Checking for mixed datatype

# Check for mixed-type data in your df_ords dataframe.

for col in df_ords.columns.tolist():
    mixeddata = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
    if len(df_ords[mixeddata]) > 0:
        print(col)

There is no mixed type data

## 8. Changing datatypes to reduce memory usage

# Change datatypes to reduce memory usage

df_ords['order_id'] = df_ords['order_id'].astype('int32')
df_ords['user_id'] = df_ords['user_id'].astype('int32')
df_ords['order_number'] = df_ords['order_number'].astype('int8')
df_ords['order_day_of_week'] = df_ords['order_day_of_week'].astype('int8')
df_ords['order_hour_of_day'] = df_ords['order_hour_of_day'].astype('int8')
df_ords['days_since_prior_order'] = df_ords['days_since_prior_order'].astype('float16')

# Checking for memory usage reduction using info

df_ords.info()

#### The shape of df_ords after consistency checks is (3421083, 6) with memory usage of 42.4 MB.

## 9. Performing Descriptive Analysis

# Checking descriptive analysis

df_ords.describe()

#### There is nothing off about the data. All the min and max values are as expected.

## 10. Exporting wrangled, consistency checked dataframe

# Export cleaned dataframe

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'), index = False)
