==> 02 Scripts/02 Products - Data Wrangling & Data Consistency Checks.ipynb <==
# Products - Data Wrangling and Data Consistency Checks:

1. Importing libraries and dataset
2. Checking for columns, datatype, shape using .info()
3. Addressing missing values
4. Addressing duplicates
5. Checking for mixed datatype
6. Changing datatypes to reduce memory usage
7. Performing Descriptive Analysis
8. Exporting wrangled, consistency checked dataframe

## 1. Importing libraries and dataset

# Importing libraries

import pandas as pd
import os

# Accessing EnvFile for path

%run EnvFile.ipynb

# Importing products_checked.csv to dataframe

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'))

# Checking the head

df_prods.head()

## 2. Checking for columns, datatype, shape using .info()

# Checking the info for columns, datatypes, shape of dataframe

df_prods.info()

#### The shape of df_prods before consistency checks is (49693, 5) with memory usage of 1.9+ MB.
#### There are no columns to be removed or renamed in this dataframe.

## 3. Addressing missing values

# Finding missing values

df_prods.isnull().sum()

# Creating a subset of missing values in product_name

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# Number of rows before removing missing values in df_prods

df_prods.shape

# Reassigning the subset of non-missing values to df_prods

df_prods = df_prods[df_prods['product_name'].isnull() == False]

# Number of rows after removing missing values in df_prods

df_prods.shape

## 4. Addressing duplicates

# Checking for duplicates

df_prods[df_prods.duplicated()]

# Reassigning the dataframe with data that doesn’t include the duplicates identified using the drop_duplicates() function

df_prods = df_prods.drop_duplicates()

# Number of rows after removing duplicate values

df_prods.shape

## 5. Checking for mixed datatype

# Check for mixed-type data in your df_ords dataframe.

for col in df_prods.columns.tolist():
    mixeddata = (df_prods[[col]].applymap(type) != df_prods[[col]].iloc[0].apply(type)).any(axis = 1)
    if len(df_prods[mixeddata]) > 0:
        print(col)

There is no mixed type data.

## 6. Changing datatypes to reduce memory usage

# Changing datatype to reduce memory usage

df_prods['product_id'] = df_prods['product_id'].astype('int32')
df_prods['aisle_id'] = df_prods['aisle_id'].astype('int8')
df_prods['department_id'] = df_prods['department_id'].astype('int8')

# Checking df info to confirm memory usage reduction

df_prods.info()

#### The shape of df_prods after consistency checks is (49672, 5) with memory usage of 1.4+ MB.

## 7. Performing Descriptive Analysis

# Checking descriptive analysis

df_prods.describe()

#### The max and std values of price is not as expected and needs to be investigated further.

# Investigate record with max 'prices' value

df_prods.loc[df_prods['prices'] == 99999]

# Check for products that cost more than $50

df_prods[df_prods["prices"]> 50.0]

# Replace 'prices' value for 'product_id' 21553 with 14.9000 and 'product_id' 33664 with 9.99  

df_prods = df_prods.replace ({"prices":{99999.0: 9.99, 14900.0:14.9 }})

# Check that changes have been successfully made

df_prods[df_prods["product_id"]==21553]

# Check that changes have been successfully made

df_prods[df_prods["product_id"]==33664]

# Rechecking descriptive analysis

df_prods.describe()

#### Now, there is nothing off about the data. All the min and max values are as expected. 

## 8. Exporting wrangled, consistency checked dataframe

# Export cleaned dataframe

df_prods.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'), index = False)
