==> 02 Scripts/03 Department - Data Wrangling and Subsetting.ipynb <==
# Departments - Data Wrangling and Subsetting:

1. Importing libraries and dataset
2. Transposing the data
3. Exporting wrangled dataframe
4. Creating Data Dictionary using .to_dict()
5. Subsetting
5.1 Snacks
5.2 Breakfast
5.3 Dinner Party

## 1. Importing libraries and dataset

# Importing libraries

import pandas as pd
import numpy as np
import os

# Accessing EnvFile for path

%run EnvFile.ipynb

# Importing products.csv

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Importing departments.csv

df_dept = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'))

df_dept.head()

## 2. Transposing the data

# Transposing df_dep

df_dept = df_dept.T

df_dept

df_dept = df_dept.reset_index()

df_dept

df_dept.info()

# Take the first row (0, department_id, department) of df_dept for the header

new_header = df_dept.iloc[0]

new_header

df_dept = df_dept[1:]

df_dept

df_dept.info()

df_dept.columns = new_header

df_dept

df_dept.info()

df_dept.reset_index(drop = True, inplace = True)

df_dept

df_dept.info()

## 3. Exporting wrangled dataframe

df_dept.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'), index = False)

## 4. Creating Data Dictionary using .to_dict()

# Creating data dictionary using .to_dict() function

dept_dict = df_dept.to_dict('index')

dept_dict

# Using .get() function in data dictionary to get the value for department_id 19

dept_dict.get('19')

## 5. Subsetting

### 5.1. The sales team in client’s organization wants to know more about snack item sales. Creating a subset containing only the required information.

# Using indexing. 
# Below code will return True/False values in df_prods based on if it matches the below condition or not.

df_prods['department_id'] == 19

# Below code will pull all the values of df_prods that match the condition within.
# Subsetting means creating a smaller data set from a whole data set (df_prods) based on a particular filter/condition.

df_prods[df_prods['department_id'] == 19]

# Above result is available while printing.
# Below code is assigning those results to a new df.

df_snacks = df_prods[df_prods['department_id'] == 19]

df_snacks.head()

### 5.2. Subset for breakfast item sales

dept_dict

df_breakfast = df_prods.loc[df_prods['department_id'] == 14]

df_breakfast

### 5.3. Client would like to see details about products that customers might use to throw dinner parties. 

### Creating subset to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood.

df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinner_parties
