==> 02 Scripts/04 Orders, Products and Department - Merging.ipynb <==
# Orders, Products and Department - Merging:
1. Importing libraries and datasets
2. Reducing memory usage and merging orders with orders_prior df
3. Reducing memory usage and merging orders_prior_merged with products df
3. Merging orders_prods_merged with dept df
4. Exporting ords_prods_dept merged df

## 1. Importing libraries and datasets

# Importing libraries

import pandas as pd
import os

# Accessing EnvFile for path

%run EnvFile.ipynb

# Importing orders_checked.csv

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'))

# Importing orders_products_prior.csv

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'))

# Importing products_checked.csv

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

df_dept = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'), index_col = False)

## 2. Reducing memory usage and merging orders with orders_prior df

df_ords.info()

df_ords_prior.info()

# Change datatypes for ords dataset to reduce memory usage

df_ords['order_id'] = df_ords['order_id'].astype('int32')
df_ords['user_id'] = df_ords['user_id'].astype('int32')
df_ords['order_number'] = df_ords['order_number'].astype('int8')
df_ords['order_day_of_week'] = df_ords['order_day_of_week'].astype('int8')
df_ords['order_hour_of_day'] = df_ords['order_hour_of_day'].astype('int8')
df_ords['days_since_prior_order'] = df_ords['days_since_prior_order'].astype('float16')

# Change datatypes for ords_prior data set 

df_ords_prior['order_id'] = df_ords_prior['order_id'].astype('int32')
df_ords_prior['product_id'] = df_ords_prior['product_id'].astype('int32')
df_ords_prior['add_to_cart_order'] = df_ords_prior['add_to_cart_order'].astype('int32')
df_ords_prior['reordered'] = df_ords_prior['reordered'].astype('int8')

df_ords.info()

df_ords_prior.info()

df_ords.head()

df_ords_prior.head()

##### In ords_prior dataset, there can be multiple rows for each order_id as this table includes the list of different products bought during each order.

# Merging df_ords and df_ords_prior based on 'order_id' using .merge() function

df_ords_prior_merged = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_ords_prior_merged.head()

# Using .value_counts() for _merge to check if there is full match or not.

df_ords_prior_merged['_merge'].value_counts()

# Dropping _merge column

df_ords_prior_merged = df_ords_prior_merged.drop(columns = ['_merge'])

# Checking the shape of merged df

df_ords_prior_merged.shape

# Export in pickle format

df_ords_prior_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_prior_merged.pkl'))

## 3. Reducing memory usage and merging orders_prior with products df

df_prods.info()

# Changing datatype to reduce memory usage

df_prods['product_id'] = df_prods['product_id'].astype('int32')
df_prods['aisle_id'] = df_prods['aisle_id'].astype('int8')
df_prods['department_id'] = df_prods['department_id'].astype('int8')

df_prods.info()

df_prods.head()

#### The prods dataset includes reference for corresponding product name and prices for each product id. So there cannot be multiple entries for each product_id.

# Checking for duplicates for product_id in prods dataset

df_prods.drop_duplicates(subset = ['product_id'])

#### There are 2 duplicates as we can notice from the shape of prods dataset reduced from 49672 to 49670. So, let's remove the duplicates and reassign the unique values to prods dataset.

df_prods = df_prods.drop_duplicates(subset = ['product_id'])

# Checking the shape of df_prods

df_prods.shape

# Merging ords_prior_merged and prods dataset

df_ords_prods_merged = df_ords_prior_merged.merge(df_prods, on = 'product_id', how = 'inner')

df_ords_prods_merged.head()

# Checking shape of merged df

df_ords_prods_merged.info()

## 3. Merging orders_prods_merged with dept df

# Merging prepared department df with current df

df_ords_prods_dept = df_ords_prods_merged.merge(df_dept, on = 'department_id')

df_ords_prods_dept.head()

# Checking memory usage after reducing

df_ords_prods_dept.info()

## 4. Exporting ords_prods_dept merged df

# Exporting orders, orders_prior and products merged df as orders_products_merged.pkl

df_ords_prods_dept.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_dept.pkl'))
