==> 02 Scripts/05 Deriving New Columns.ipynb <==
# Deriving New Columns:

1. Importing libraries and dataset
2. Creating column for price range using .loc()
3. Creating column to categorise day busyness level over a week using for and if else condition
4. Creating column to categorise hour busyness level over a day using for and if else condition
5. Exporting the dataset with newly derived columns

## 1. Importing libraries and dataset

# Importing libraries

import pandas as pd
import numpy as np
import os

# Accessing EnvFile for path

%run EnvFile.ipynb

# Importing orders_products_merged.pkl

df_ords_prods_dept = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_dept.pkl'))

df_ords_prods_dept.info()

## 2. Creating column for price range using .loc()

df_ords_prods_dept.loc[df_ords_prods_dept['prices'] > 15, 'price_range'] = 'High-range product'

df_ords_prods_dept.loc[(df_ords_prods_dept['prices'] <= 15) & (df_ords_prods_dept['prices'] > 5), 'price_range'] = 'Mid-range product'

df_ords_prods_dept.loc[df_ords_prods_dept['prices'] <= 5, 'price_range'] = 'Low-range product'

df_ords_prods_dept['price_range'].value_counts(dropna = False)

## 3. Creating column to categorise day busyness level over the week

df_ords_prods_dept['order_day_of_week'].value_counts(dropna = False)

#### 0 (Saturday) and 1 (Sunday) are the most busy days, 4 (Wednesday) and 3 (Tuesday) are the least busy days and the rest are regular busy days.

day_busyness_level = []
for value in df_ords_prods_dept['order_day_of_week']:
    if (value == 0) or (value == 1):
        day_busyness_level.append('Most Busy Day')
    elif (value == 4) or (value == 3):
        day_busyness_level.append('Least Busy Day')
    else:
        day_busyness_level.append('Average Busy Day')

# Assigning the above for loop result to a new column in df

df_ords_prods_dept['day_busyness_level'] = day_busyness_level

# Finding the frequency of new column

df_ords_prods_dept['day_busyness_level'].value_counts(dropna = False)

## 4. Creating column to categorise hour busyness level over the day

df_ords_prods_dept['order_hour_of_day'].value_counts(dropna = False)

hour_busyness_level = []
for value in df_ords_prods_dept['order_hour_of_day']:
    if (value == 10) or (value == 11):
        hour_busyness_level.append('Most Busy Hour')
    elif (value == 3) or (value == 4):
        hour_busyness_level.append('Least Busy Hour')
    else:
        hour_busyness_level.append('Average Busy Hour')

# Assigning the above for loop result to new column in df

df_ords_prods_dept['hour_busyness_level'] = hour_busyness_level

# Finding the frequency of newly created column

df_ords_prods_dept['hour_busyness_level'].value_counts(dropna = False)

# Checking the updated df with 3 newly created columns

df_ords_prods_dept.head()

df_ords_prods_dept.info()

## 5. Exporting the dataset with newly dervied columns

# Exporting the dataframe with new variables

df_ords_prods_dept.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_dept_newcolumns.pkl'))
