==> 02 Scripts/06 Grouping Data & Aggregating Variables.ipynb <==
# Grouping Data & Aggregating Variables:

1. Importing libraries and dataset
2. Finding the aggregated mean of the “order_number” column grouped by “department_id”
3. Creating a loyalty flag for existing customers using the transform() and loc() functions
4. Exploring spending habits of customer for marketing team
5. Categorising the different types of spenders to target via marketing campaigns
6. Determine frequent versus non-frequent customers in order to send relevant notifications to users within the app
7. Exporting the dataframe

## 1. Importing libraries and dataset

# Importing libraries

import pandas as pd
import numpy as np
import os

# Accessing EnvFile for path

%run EnvFile.ipynb

# Importing dataframe

df_ords_prods_dept = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_dept_newcolumns.pkl'))

## 2. Finding the aggregated mean of the “order_number” column grouped by “department_id”

df_ords_prods_dept.groupby('department_id').agg({'order_number' : ['mean']})

## 3. Creating a loyalty flag for existing customers using the transform() and loc() functions

# Split the data (user_id) into groups by .groupby() and assign new column (max_order) with aggregate values (max) of order_number using .transform()

df_ords_prods_dept['max_order'] = df_ords_prods_dept.groupby(['user_id'])['order_number'].transform(np.max)

# Checking the newly created column using index and head()

df_ords_prods_dept[['user_id', 'order_number', 'max_order']].head(60)

# Creating loyal customer condition using .loc()

df_ords_prods_dept.loc[df_ords_prods_dept['max_order'] > 40, 'loyalty_flag'] = 'Loyal Customer'

# Creating regular customer condition using .loc()

df_ords_prods_dept.loc[(df_ords_prods_dept['max_order'] <= 40) & (df_ords_prods_dept['max_order'] > 10), 'loyalty_flag'] = 'Regular Customer'

# Creating new customer condition using .loc()

df_ords_prods_dept.loc[df_ords_prods_dept['max_order'] <= 10, 'loyalty_flag'] = 'New Customer'

# Checking the frequency of new flag column using .value_counts()

df_ords_prods_dept['loyalty_flag'].value_counts(dropna = False)

# Checking the specific column results using index and head()

df_ords_prods_dept[['user_id', 'order_number', 'max_order', 'loyalty_flag']].head(15)

## 4. Exploring spending habits of customer for marketing team

# Split the data (loyalty_flag) into groups using groupby() and apply aggregate functions using agg() to obtain basic statistics (mean, min, max values) of prices

df_ords_prods_dept.groupby('loyalty_flag').agg({'prices' : ['mean', 'min', 'max']})

#### The prices of products purchased by loyal customers do not differ hugely from those purchased by regular or new customers.

## 5. Categorising the different types of spenders to target via marketing campaigns

Creating a spending flag for each user based on the average price across all their orders using the following criteria:

- If the mean of the prices of products purchased by a user is lower than 10, then flagging them as a “Low Spender.”
- If the mean of the prices of products purchased by a user is higher than or equal to 10, then flagging them as a “High Spender.”

# Split the data (user_id) into groups by .groupby() and assign new column (mean_price) with aggregate values (max) of prices using .transform()

df_ords_prods_dept['average_price'] = df_ords_prods_dept.groupby(['user_id'])['prices'].transform(np.mean)

# Checking the newly created column using .head()

df_ords_prods_dept[['user_id', 'prices', 'average_price']].head(15)

# Creating low spender condition using .loc()

df_ords_prods_dept.loc[df_ords_prods_dept['average_price'] < 10, 'spending_flag'] = 'Low Spender'

# Creating high spender condition using .loc()

df_ords_prods_dept.loc[df_ords_prods_dept['average_price'] >= 10, 'spending_flag'] = 'High Spender'

# Checking the frequency of new flag column using .value_counts()

df_ords_prods_dept['spending_flag'].value_counts(dropna = False)

# Checking the specific column results using index and head()

df_ords_prods_dept[['user_id', 'prices', 'average_price', 'spending_flag']].head(15)

## 6. Determine frequent versus non-frequent customers in order to send relevant notifications to users within the app

Creating an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. The criteria for the flag will be as follows:
- If the median of “days_since_prior_order” is higher than 20, then the customer will be labeled a “Non-frequent Customer.”
- If the median is higher than 10 and lower than or equal to 20, then the customer will be labeled a “Regular Customer.”
- If the median is lower than or equal to 10, then the customer will be labeled a “Frequent Customer.”

# Split the data (user_id) into groups by .groupby() and assign new column (median_days_since_prior_order) with aggregate values (median) of days_since_prior_order using .transform()

df_ords_prods_dept['median_days_since_prior_order'] = df_ords_prods_dept.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# Checking the newly created column using index and .head()

df_ords_prods_dept[['user_id', 'days_since_prior_order', 'median_days_since_prior_order']].head(15)

# Creating non-frequent customer condition using .loc()

df_ords_prods_dept.loc[df_ords_prods_dept['median_days_since_prior_order'] > 20, 'order_frequency_flag'] = 'Non-frequent Customer'

# Creating regular customer condition using .loc()

df_ords_prods_dept.loc[(df_ords_prods_dept['median_days_since_prior_order'] > 10) & (df_ords_prods_dept['median_days_since_prior_order'] <= 20), 'order_frequency_flag'] = 'Regular Customer'

# Creating frequent customer condition using .loc()

df_ords_prods_dept.loc[df_ords_prods_dept['median_days_since_prior_order'] <= 10, 'order_frequency_flag'] = 'Frequent Customer'

# Checking the frequency of new flag column using .value_counts()

df_ords_prods_dept['order_frequency_flag'].value_counts(dropna = False)

# Checking the specific column results using index and head()

df_ords_prods_dept[['user_id', 'days_since_prior_order', 'median_days_since_prior_order', 'order_frequency_flag']].head(15)

## 7. Exporting the dataframe

# Exporting the updated df with new aggregated columns

df_ords_prods_dept.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_dept_aggcolumns.pkl'))
